added ChallengeAndUser collection+gototask button.

+++ lib/DatabaseModels/ChallengeAndUser.dart
@@ -6,9 +6,9 @@ class ChallengeAndUser {
   double? currentryRaised;
   String? challengeType;
   String? leetcodeUsername;
-  double? easy;
-  double? medium;
-  double? hard;
+  int? easy;
+  int? medium;
+  int? hard;
   bool? exited;
   DateTime? createdAt;
 
@@ -33,9 +33,9 @@ class ChallengeAndUser {
         currentryRaised = snapshot.data()['currentryRaised'] as double,
         challengeType = snapshot.data()['challengeType'],
         leetcodeUsername = snapshot.data()['leetcodeUsername'],
-        easy = snapshot.data()['easy'] as double,
-        medium = snapshot.data()['medium'] as double,
-        hard = snapshot.data()['hard'] as double,
+        easy = snapshot.data()['easy'] as int,
+        medium = snapshot.data()['medium'] as int,
+        hard = snapshot.data()['hard'] as int,
         exited = snapshot.data()['exited'],
         createdAt = snapshot.data()['createdAt'].toDate();
 }
